src: change FormatSize to actually accept a size_t
This function is only called with size_t values and it does not deal
with unsigned values nicely, so its argument type should be size_t.

PR-URL: https://github.com/nodejs/node/pull/43464

+++ src/node_snapshotable.cc
@@ -57,7 +57,7 @@ static std::string GetCodeCacheDefName(const std::string& id) {
   return std::string(buf) + std::string("_cache_data");
 }
 
-static std::string FormatSize(int size) {
+static std::string FormatSize(size_t size) {
   char buf[64] = {0};
   if (size < 1024) {
     snprintf(buf, sizeof(buf), "%.2fB", static_cast<double>(size));
